# speed up new block ntfn and storage handling