@chappjc 2019-10-27
Indexing the `blocks` table on `"time"` seems to be the silver bullet.

On a not-so-fast production VM:

**Before**

```
[DBG] NTFN: handlers of Notifier.processBlock() completed in 11.979038937s
```

**After**

```
[DBG] NTFN: handlers of Notifier.processBlock() completed in 323.007192ms
```

A DB scheme upgrade is generally a last resort, but this is just a new index, and the speed up is dramatic.