@chappjc 2019-10-27
I'm tempted to backport this to 5.1 since `Notifier.processBlock()` is unacceptably slow, but I think it's too much.  We'll just have to get 5.2 out the door.