@chappjc 2019-10-28
Been running stable on alpha for >12hrs now.  `Notifier.processBlock` exec time reduced from ~12sec to <1sec.